+++ api/src/main/java/io/strimzi/api/kafka/model/Kafka.java
@@ -48,7 +48,8 @@ import static java.util.Collections.unmodifiableList;
                         @Crd.Spec.Version(name = Kafka.V1ALPHA1, served = true, storage = false)
                 },
                 subresources = @Crd.Spec.Subresources(
-                               status = @Crd.Spec.Subresources.Status()
+                        status = @Crd.Spec.Subresources.Status(),
+                        scale = {}
                 ),
                 additionalPrinterColumns = {
                         @Crd.Spec.AdditionalPrinterColumn(