+++ api/src/main/java/io/strimzi/api/kafka/model/KafkaConnectS2I.java
@@ -59,7 +59,12 @@ import static java.util.Collections.unmodifiableList;
                         )
                 },
                 subresources = @Crd.Spec.Subresources(
-                        status = @Crd.Spec.Subresources.Status()
+                        status = @Crd.Spec.Subresources.Status(),
+                        scale = @Crd.Spec.Subresources.Scale(
+                                specReplicasPath = KafkaConnectS2I.SPEC_REPLICAS_PATH,
+                                statusReplicasPath = KafkaConnectS2I.STATUS_REPLICAS_PATH,
+                                labelSelectorPath = KafkaConnectS2I.LABEL_SELECTOR_PATH
+                        )
                 ),
                 additionalPrinterColumns = {
                         @Crd.Spec.AdditionalPrinterColumn(
@@ -96,6 +101,9 @@ public class KafkaConnectS2I extends CustomResource implements UnknownPropertyPr
     public static final String CRD_NAME = RESOURCE_PLURAL + "." + RESOURCE_GROUP;
     public static final String SHORT_NAME = "kcs2i";
     public static final List<String> RESOURCE_SHORTNAMES = singletonList(SHORT_NAME);
+    public static final String SPEC_REPLICAS_PATH = ".spec.replicas";
+    public static final String STATUS_REPLICAS_PATH = ".status.replicas";
+    public static final String LABEL_SELECTOR_PATH = ".status.selector";
 
     private String apiVersion;
     private ObjectMeta metadata;