+++ api/src/main/java/io/strimzi/api/kafka/model/KafkaMirrorMaker2.java
@@ -50,7 +50,12 @@ import static java.util.Collections.unmodifiableList;
                         )
                 },
                 subresources = @Crd.Spec.Subresources(
-                        status = @Crd.Spec.Subresources.Status()
+                        status = @Crd.Spec.Subresources.Status(),
+                        scale = @Crd.Spec.Subresources.Scale(
+                                specReplicasPath = KafkaMirrorMaker2.SPEC_REPLICAS_PATH,
+                                statusReplicasPath = KafkaMirrorMaker2.STATUS_REPLICAS_PATH,
+                                labelSelectorPath = KafkaMirrorMaker2.LABEL_SELECTOR_PATH
+                        )
                 ),
                 additionalPrinterColumns = {
                         @Crd.Spec.AdditionalPrinterColumn(
@@ -87,6 +92,9 @@ public class KafkaMirrorMaker2 extends CustomResource implements UnknownProperty
     public static final String CRD_NAME = RESOURCE_PLURAL + "." + RESOURCE_GROUP;
     public static final String SHORT_NAME = "kmm2";
     public static final List<String> RESOURCE_SHORTNAMES = singletonList(SHORT_NAME);
+    public static final String SPEC_REPLICAS_PATH = ".spec.replicas";
+    public static final String STATUS_REPLICAS_PATH = ".status.replicas";
+    public static final String LABEL_SELECTOR_PATH = ".status.selector";
 
     private String apiVersion;
     private KafkaMirrorMaker2Spec spec;