+++ api/src/main/java/io/strimzi/api/kafka/model/KafkaRebalance.java
@@ -50,7 +50,8 @@ import static java.util.Collections.unmodifiableList;
                         )
                 },
                 subresources = @Crd.Spec.Subresources(
-                        status = @Crd.Spec.Subresources.Status()
+                        status = @Crd.Spec.Subresources.Status(),
+                        scale = {}
                 )
         )
 )