+++ crd-generator/src/main/java/io/strimzi/crdgenerator/annotations/Crd.java
@@ -4,6 +4,8 @@
  */
 package io.strimzi.crdgenerator.annotations;
 
+import io.fabric8.kubernetes.api.model.extensions.Scale;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -108,7 +110,10 @@ public @interface Crd {
          * @return The subresources of a custom resources that this is the definition for.
          * @see <a href="https://v1-11.docs.kubernetes.io/docs/reference/generated/kubernetes-api/v1.11/#customresourcedefinitionversion-v1beta1-apiextensions">Kubernetes 1.11 API documtation</a>
          */
-        Subresources subresources() default @Subresources(status = {});
+        Subresources subresources() default @Subresources(
+                status = {},
+                scale = {}
+                );
 
         /**
          * The subresources of a custom resources that this is the definition for.
@@ -116,8 +121,19 @@ public @interface Crd {
          */
         @interface Subresources {
             Status[] status();
+            Scale[] scale();
 
             @interface Status {
+            }
+
+            /**
+             * The scale subresource of a custom resources that this is the definition for.
+             * @see <a href="https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.18/#customresourcesubresourcescale-v1beta1-apiextensions-k8s-io">Kubernetes 1.18 API documtation</a>
+             */
+            @interface Scale {
+                String specReplicasPath();
+                String statusReplicasPath();
+                String labelSelectorPath() default "";
             }
         }